pyproject.toml
[tool.poetry]
name = "incandescent-bot"
version = "0.14.0"
description = "A Discord bot for Incandescent Society Discord server written in Python."
authors = ["kyomi <me@kyomi.dev>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
discord-py = "^2.3.1"
jishaku = "^2.5.1"
sqlalchemy = {version = "^2.0.17", extras = ["asyncio"]}
alembic = "^1.11.1"
asyncpg = "^0.27.0"
psutil = "^5.9.5"
humanize = "^4.7.0"
aiocron = "^1.8"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.3.0"
isort = "^5.12.0"
commitizen = "^3.5.2"
flake8 = "^6.0.0"
pyright = "^1.1.316"

[tool.black]
color = true
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.commitizen]
version = "0.14.0"
version_files = ["pyproject.toml"]
update_changelog_on_bump = true
changelog_incremental = true

[tool.pyright]
typeCheckingMode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
